Set a GITHUB_TOKEN env. var in checker pipeline
Version 2.11 of the orphaned namespace checker uses a github
personal access token, so that it doesn't run into API rate limits
for unauthenticated github API calls.

This commit tells the checker pipeline to use version 2.11 of the
docker image, and exposes an existing github access token (which
Vijay created for the PR part of the plan pipeline) as a
GITHUB_TOKEN environment variable.