Reconcile AWS IAM policy with what's in moj-cp
A `tf plan` from the master branch of this repo, shows that it
wants to make the following changes to the IAM AWS policy in moj-cp.

              -        "route53:ChangeTagsForResource",
              -        "route53:ChangeResourceRecordSets",
              -        "route53:ListResourceRecordSets"
              +        "route53:ChangeTagsForResource"

This indicates that someone manually added ChangeResourceRecordSets
and ListResourceRecordSets to the AWS policy without updating the
terraform source code.

This commit brings the terraform source into line with the current
AWS policy in the moj-cp account.

However, the platforms-integration account has not been updated.
So, when this PR is merged, and applied using terraform, it will
make changes to the policy on the platforms-integration AWS account
as follows:

              +        "route53:ListResourceRecordSets",
                       "route53:GetHostedZone",
                       "route53:DeleteHostedZone",
              -        "route53:ChangeTagsForResource"
              +        "route53:ChangeTagsForResource",
              +        "route53:ChangeResourceRecordSets"

IMO, this should be fine. We're only adding permissions, not taking
any away, so there shouldn't be any visible effect from this change.