Tag the tools image as 1.4
We were using 1.4-test to see if changing the image to CMD from
ENTRYPOINT caused any problems. It doesn't seem as if it does, so
we can now build and use 1.4 everywhere.

I've manually built and pushed the 1.4 tag to docker hub already,
so it should be fine to specify 1.4 in the plan and build
pipelines now, rather than having separate PRs for tagging the
image and using the image.