Add TF_VAR_xxx environment variables to live-1
This change should make no difference to the 
current live-1 pipelines, but it is a pre-
requisite for adapting the pipeline code to work
with terraform 0.12 source code.

Terraform 0.12 won't let you pass a variable on
the command-line unless the terraform source code
uses the variable, but this restriction does not
apply to TF_VAR_xxx environment variables.