Add a new pipeline to detect deleted namespaces
Currently, the script will just post a slack message to tell us
to delete the relevant namespaces. As such, we don't need all the
environment variables configured here. But, once we're happy that
the script works correctly, we will modify it to auto-delete the
namespaces, at which point it *will* need all of these env. vars.
So, I've added all the env vars here, so that a later change
should only need to update the ruby script in the environments
repo, and won't need to change this pipeline.