Remove unused env. vars for namespace deleter
depends on
https://github.com/ministryofjustice/cloud-platform-environments/pull/1733

Previously, the code this was based upon had to
work on both live-0 and live-1, which are in
different AWS accounts, so in one cluster we
needed two different sets of AWS credentials.

Now that we only have to worry about live-1, we
can be sure that the same AWS credentials that
give us access to the terraform state files will
also give us access to the kubeconfig file's S3
bucket. So, we can reduce complexity by only
using a single set of AWS credentials.