Add a github access token as a secret
This puts a github personal access token for the
"how out of date are we" updater concourse
pipeline into the concourse-main namespace on the
live-1 and manager clusters.

It is only necessary to put this secret into the
manager cluster, since that is where the pipeline
runs, but I wasn't sure if only defining the 
secret for one cluster would cause problems when
the terraform code is applied, so I added it to
both.