Run the HOODAW updater every hour
The HOODAW updater works perfectly when I trigger the pipeline via the
concourse web interface, but although it was running on its 24h
schedule, the application did not show that the data had been
successfully submitted.

This change switches to an hourly schedule, so that we get a clearer
sense of whether or not the pipeline/updater are actually working
properly or not.