Update the HOODAW github token
HOODAW needs a github personal access token in 
order to query the github graphql API for repo.
metadata.

Prior to this change, the token belonged to the
`cloud-platform-moj` 'bot' github account. But, 
this token does not (and cannot) have MoJ SSO
enabled, which prevents it from retrieving the
data we need.

We are in the process of getting a "real" google
account to attach to this github account, so that
we can log it in via SSO, and then enable SSO on
the github token.

In the meantime, this change sets the github token
to one belonging to my github account 
`digitalronin`, which has SSO enabled.

THIS WILL STOP WORKING WHEN I LEAVE THE MOJ

So, this is just an interim band-aid, until we can
fix the problem properly (which depends on the
digitalservicedesk creating the google account)