Remove the "build tools image" pipeline
We now have a github action to build and push the tools image to docker
hub whenever we create a new release, so this pipeline is redundant.

It's also incorrect because we haven't been updating the tag value, so
"1.15" has been continually rebuilt at later and later versions of the
tools image.