Remove user guide repo from the exceptions list
The user guide repo still *has* a master branch, 
in order to publish master/docs via github pages.
But the master branch is no longer the default
branch, and is not subject to branch protection.
So it doesn't need to be treated as an exception
anymore.

related to https://github.com/ministryofjustice/cloud-platform-user-guide/pull/339