Change terraform log-level to normal
We increased the terraform log level to DEBUG for the cost reporter pipeline, while tracking down an intermittent problem affecting this pipeline and the main "apply" pipeline. The problem was tracked down to the Pingdom API sometimes failing to give a valid response, so this extra logging is no longer necessary.