Pass HOODAW_API_KEY env. var. to HOODAW updater
The updater used to fetch the HOODAW api key directly from the
secret in the how-out-of-date-are-we namespace. However, we now
have the api key explicitly set as a secret available to concourse.

See
https://github.com/ministryofjustice/cloud-platform-terraform-concourse/pull/9

So, we can pass that value to the updater image.

A related PR will remove the code that fetches the env. var from the
namespace, so that we always get this credential in the same way, in all
pipeline jobs.