Run namespace cost report every 12 hours
This report currently takes ~2 hours to run, but it often fails partway through. I think this is because of contention with the apply pipeline, which also locks the terraform state of each namespace in turn.
This change runs the namespace cost report every 12 hours, instead of every 24, so that the HOODAW namespace cost report will be more up to date.
The report randomises the order in which it reports namespaces, so eventually we will get all the data, even if there is a failure on every run.