Use v2.25 of the orphaned-namespace-checker image
This image runs with a non-root user, so for the reporting pipeline,
which passes output from one step as input to another, we need to tell
concourse to override the image and run the job as root. Without that,
the job fails with permission errors. I tried pre-creating and chowning
various permutations of directories, but in the end specifying `root` as
the task executor was the only way I managed to get it working.