Run the orphaned namespace checker daily
If someone deletes a production namespace via a PR 
in the env. repo, the 'delete namespace' pipeline 
will fail with a message that it won't delete prod 
namespaces.

This is easy to miss, especially if 
#lower-priority-alerts is noisy.

Also, if an additional, non-production namespace 
is deleted at the same time, the pipeline might 
not notice that one, if it dies because of the 
production namespace.

Running the orphaned namespace checker every day
means we should notice such issues sooner.